# backend/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Open Chat application",
        "title": "Open Chat API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/integrations/{integration_name}/overview": {
            "get": {
                "description": "Returns API routes, registered model types and function names for a compiled integration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Get integration overview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Integration name",
                        "name": "integration_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/integrations.IntegrationOverviewResponse"
                        }
                    },
                    "404": {
                        "description": "integration not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/register": {
            "post": {
                "description": "Register a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Register a user",
                "parameters": [
                    {
                        "description": "User registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRegister"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Password too short",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "integrations.IntegrationAPIParameterOverview": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "in": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "integrations.IntegrationAPIRouteOverview": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationAPIParameterOverview"
                    }
                },
                "required_auth": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "route": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "integrations.IntegrationFrontendRouteOverview": {
            "type": "object",
            "properties": {
                "asset_path": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "route": {
                    "type": "string"
                }
            }
        },
...
        "integrations.IntegrationModelFieldOverview": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationModelFieldOverview"
                    }
                },
                "json_name": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "integrations.IntegrationModelOverview": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationModelFieldOverview"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
        "integrations.IntegrationOverviewResponse": {
            "type": "object",
            "properties": {
                "api_routes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "api_routes_overview": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationAPIRouteOverview"
                    }
                },
                "frontend_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationFrontendRouteOverview"
                    }
                },
                "functions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationModelOverview"
                    }
                },
                "name": {
                    "type": "string"
                },
                "readme_markdown": {
                    "type": "string"
                },
                "runtime_env_vars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/integrations.IntegrationRuntimeEnvVarOverview"
                    }
                }
            }
        },
        "integrations.IntegrationRuntimeEnvVarOverview": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "sensitive": {
                    "type": "boolean"
                }
            }
        },
...
        "user.UserRegister": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie obtained from login endpoint",
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    },
...
}